+++ substratevm/src/com.oracle.graal.pointsto/src/com/oracle/graal/pointsto/heap/HostedValuesProvider.java
@@ -26,6 +26,7 @@ package com.oracle.graal.pointsto.heap;
 
 import com.oracle.graal.pointsto.heap.value.ValueSupplier;
 import com.oracle.graal.pointsto.meta.AnalysisField;
+import com.oracle.graal.pointsto.meta.AnalysisMetaAccess;
 import com.oracle.graal.pointsto.meta.AnalysisUniverse;
 import com.oracle.graal.pointsto.util.GraalAccess;
 
@@ -33,9 +34,11 @@ import jdk.vm.ci.meta.JavaConstant;
 import jdk.vm.ci.meta.JavaKind;
 
 public class HostedValuesProvider {
+    protected final AnalysisMetaAccess metaAccess;
     protected final AnalysisUniverse universe;
 
-    public HostedValuesProvider(AnalysisUniverse universe) {
+    public HostedValuesProvider(AnalysisMetaAccess metaAccess, AnalysisUniverse universe) {
+        this.metaAccess = metaAccess;
         this.universe = universe;
     }