+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/meta/RelocatableConstant.java
@@ -26,19 +26,21 @@ package com.oracle.svm.hosted.meta;
 
 import org.graalvm.nativeimage.c.function.RelocatedPointer;
 
-import jdk.graal.compiler.core.common.type.TypedConstant;
+import com.oracle.graal.pointsto.heap.TypedConstant;
+import com.oracle.graal.pointsto.meta.AnalysisType;
+
 import jdk.vm.ci.meta.JavaConstant;
 import jdk.vm.ci.meta.JavaKind;
-import jdk.vm.ci.meta.MetaAccessProvider;
-import jdk.vm.ci.meta.ResolvedJavaType;
 
 /** Wraps pointers that are subject to relocation, so their value is not known during analysis. */
 public class RelocatableConstant implements JavaConstant, TypedConstant {
 
     private final RelocatedPointer pointer;
+    private final AnalysisType type;
 
-    public RelocatableConstant(RelocatedPointer pointer) {
+    public RelocatableConstant(RelocatedPointer pointer, AnalysisType type) {
         this.pointer = pointer;
+        this.type = type;
     }
 
     public RelocatedPointer getPointer() {
@@ -91,8 +93,8 @@ public class RelocatableConstant implements JavaConstant, TypedConstant {
     }
 
     @Override
-    public ResolvedJavaType getType(MetaAccessProvider provider) {
-        return provider.lookupJavaType(pointer.getClass());
+    public AnalysisType getType() {
+        return type;
     }
 
     @Override